+++ server/src/index.ts
@@ -2,74 +2,35 @@ import express, { Express, Request, Response } from 'express';
 import dotenv from 'dotenv';
 import mongoose from 'mongoose';
 import cors from 'cors';
-import Article from './models/Article';
+import Article from './models/Article.js';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+import articleRouter from './routes/articleRouter.js';
 
 dotenv.config();
 
+const MONGO_DB_URL = process.env.MONGO_DB_URL || "";
 const app: Express = express();
 const port = process.env.PORT || 3000;
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 app.use(cors());
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
+app.use("/api/article", articleRouter);
 
 mongoose
-  .connect('mongodb://localhost:27017/mydatabase')
+  .connect(MONGO_DB_URL)
   .then(() => console.log('mongoDB connected'))
-  .catch((err) => console.log(err));
+  .catch((err: Error) => console.log(err));
 
 const db = mongoose.connection;
 db.on('error', console.error.bind(console, 'connection error:'));
 db.once('open', () => {
   console.log('conected succesfully');
   initializeDb();
-});
-
-app.get('/api/article/:articleId', async (req: Request, res: Response) => {
-  try {
-    const { articleId } = req.params;
-    const articleIdNumber = Number(articleId);
-
-    if (isNaN(articleIdNumber)) return res.status(400).json({ message: 'Invalid article ID' });
-
-    const article = await Article.findOne({ id: articleIdNumber });
-
-    if (!article) {
-      return res.status(404).json({ message: 'Article not found' });
-    }
-
-    res.json(article);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ message: 'Server error' });
-  }
-});
-
-app.patch('/api/article/:articleId', async (req: Request, res: Response) => {
-  try {
-    const { articleId } = req.params;
-    const articleIdNumber = Number(articleId);
-    console.log(req.body);
-    const { content } = req.body;
-
-    const updatedArticle = await Article.findOneAndUpdate(
-      { id: articleIdNumber },
-      {
-        content,
-        updatedAt: new Date(),
-      },
-      { new: true, runValidators: true }
-    );
-
-    if (!updatedArticle) {
-      return res.status(404).json({ message: 'Article not found' });
-    }
-
-    res.json(updatedArticle);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ message: 'Server error' });
-  }
 });
 
 app.listen(port, () => {
